requirements.txt
Django>=6.0,<7.0
psycopg2-binary>=2.9,<3
dj-database-url>=3.1,<4
whitenoise>=6.7,<7
gunicorn>=21,<26
djangorestframework>=3.15,<4
django-filter>=25.0,<26
matplotlib>=3.6,<4
numpy>=1.26,<3
pandas>=2.1,<4
